requirements.txt
python-dotenv==1.2.2
Flask==3.0.3
Pillow==10.4.0
numpy==2.1.1
scikit-learn==1.5.2
pycolornames==1.0.2
gunicorn==23.0.0